voilivoilou

+++ Brevaide/src/data/brevaide/QCM.java
@@ -5,7 +5,7 @@ import java.util.Date;
 
 
 public class QCM {
-	public long id;
+	public int id;
 	public Date date;
 	public String matiere;
 	public int score;
@@ -26,10 +26,10 @@ public class QCM {
 
 
 
-	public long getId() {
+	public int getId() {
 		return id;
 	}
-	public void setId(long id) {
+	public void setId(int id) {
 		this.id = id;
 	}
